# api.yaml
openapi: 3.0.0
info:
  title: Notes API
  description: API for managing notes and users
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local server

paths:
  /notes:
    get:
      summary: Get all notes
      responses:
        '200':
          description: A list of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    user_id:
                      type: integer
                    title:
                      type: string
                    contents:
                      type: string
    post:
      summary: Create a new note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                title:
                  type: string
                contents:
                  type: string
      responses:
        '201':
          description: Note created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  title:
                    type: string
                  contents:
                    type: string

  /note/{id}:
    get:
      summary: Get a note by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the note
          schema:
            type: integer
      responses:
        '200':
          description: The note
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  title:
                    type: string
                  contents:
                    type: string
        '404':
          description: Note not found

  /note/{title}:
    get:
      summary: Get a note by its title
      parameters:
        - name: title
          in: path
          required: true
          description: The title of the note
          schema:
            type: string
      responses:
        '200':
          description: The note
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  title:
                    type: string
                  contents:
                    type: string
        '404':
          description: Note not found

  /users:
    get:
      summary: Get the list of users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    username:
                      type: string
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string

  /user/{id}:
    get:
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the user
          schema:
            type: integer
      responses:
        '200':
          description: The user
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
        '404':
          description: User not found

  /post/{id}:
    delete:
      summary: Delete a post by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the post
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Post deleted successfully
        '404':
          description: Post not found

  /post:
    patch:
      summary: Update an existing post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                content:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Post updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  title:
                    type: string
                  contents:
                    type: string

  /user:
    patch:
      summary: Update user information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                username:
                  type: string
                old_username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string

  /about:
    get:
      summary: Get API documentation
      responses:
        '200':
          description: Swagger UI documentation